fix: avoid frontend env dependency
