+++ .env.example
@@ -1,20 +1,3 @@
-# Google Sheets設定
-GOOGLE_SHEETS_ID=your_spreadsheet_id_here
-GOOGLE_SHEETS_NAME=修正用シート
-
-# Google Service Account認証情報
-# Google Cloud ConsoleでService Accountを作成し、JSONキーをダウンロード
-# その内容を1行にして貼り付ける
-GOOGLE_SERVICE_ACCOUNT_KEY={"type":"service_account","project_id":"..."}
-
-# Google OAuth認証（オプション - ユーザー認証を使う場合）
-GOOGLE_CLIENT_ID=your_google_client_id
-GOOGLE_CLIENT_SECRET=your_google_client_secret
-
-# NextAuth設定
-NEXTAUTH_URL=http://localhost:3000
-NEXTAUTH_SECRET=your_random_secret_key_here
-
-# アプリケーション設定
-NEXT_PUBLIC_APP_URL=http://localhost:3000
-NEXT_PUBLIC_APP_NAME=梱包日報管理システム
+NEXT_PUBLIC_GAS_ENDPOINT=
+NEXT_PUBLIC_SHEET_ID=
+NEXT_PUBLIC_SHEET_NAME=修正用シート
