--- .env.local
@@ -1,3 +1,0 @@
-# .env.local
-NEXT_PUBLIC_GAS_WEBAPP_URL=https://script.google.com/macros/s/AKfycbyC-AclhbsO5v4tGp7KN3Mm4RSoLM8jb2z9Ih_LIuqGECeqI_vhRNES53KxC3oJ3_k2CQ/exec
-NEXT_PUBLIC_ENVIRONMENT=TEST
