+++ pages/api/packing/search.ts
@@ -10,6 +10,14 @@ export default async function handler(
     return res.status(405).json({ error: 'Method not allowed' });
   }
 
+  const GAS_ENDPOINT = process.env.NEXT_PUBLIC_GAS_ENDPOINT;
+  if (!GAS_ENDPOINT) {
+    return res.status(500).json({
+      success: false,
+      error: 'GAS endpoint is not configured'
+    });
+  }
+
   try {
     const { date, product, status, quantityMin, quantityMax } = req.query;
